Add & Refactor Report

+++ app/schools/views.py
@@ -2,6 +2,8 @@ from django.shortcuts import render, get_object_or_404, redirect
 
 from .forms import CommentForm, ReplyForm
 from .models import School, Comment, Reply
+from ..reports.forms import ReportForm
+from ..reports.views import report_comment
 from ..users.models import UserBase
 
 
@@ -17,11 +19,21 @@ def school_list(request):
 def school_detail(request, pk):
     school = get_object_or_404(School, pk=pk)
     comments = Comment.objects.filter(school=school).order_by('-created_at')
+    report_forms = {}
+    for comment in comments:
+        form = ReportForm()
+        form.fields['reason'].widget.attrs.update({
+            'id': f'reason-{comment.id}',
+            'class': 'form-control',
+            'placeholder': 'Explica la razón del reporte...'
+        })
+        report_forms[comment.id] = form
 
     if request.method == "GET":
         context = {
             'school': school,
-            'comments': comments
+            'comments': comments,
+            'report_forms': report_forms,
         }
 
         return render(request, 'school/school_detail.html', context)
@@ -54,9 +66,18 @@ def edit_comment(request, pk):
         form = CommentForm(instance=comment)
     return render(request, 'edit_comment.html', {'form': form})
 
-
 def delete_comment(request, pk, idComentario):
     comment = get_object_or_404(Comment, pk=idComentario)
+    if not request.user.is_authenticated:
+        return render(request, 'school/school_detail.html', {
+            'school': comment.school,
+            'comments': Comment.objects.filter(school=comment.school),
+        })
+    if not (request.user.is_superuser or request.user.is_staff):
+        return render(request, 'school/school_detail.html', {
+            'school': comment.school,
+            'comments': Comment.objects.filter(school=comment.school),
+        })
     comment.delete()
     return redirect('school:school_detail', pk=pk)
 
@@ -77,6 +98,16 @@ def edit_reply(request, pk):
 
 def delete_reply(request, pk, idComentario, idRespuesta):
     reply = get_object_or_404(Reply, pk=idRespuesta)
+    if not request.user.is_authenticated:
+        return render(request, 'school/school_detail.html', {
+            'school': reply.school,
+            'comments': Comment.objects.filter(school=reply.school),
+        })
+    if not (request.user.is_superuser or request.user.is_staff):
+        return render(request, 'school/school_detail.html', {
+            'school': reply.school,
+            'comments': Comment.objects.filter(school=reply.school),
+        })
     reply.delete()
     return redirect('school:school_detail', pk=pk)
